Initial modularization
allow for built-in modules to support nightmare. This refactor separates
cookies into it's own submodule
