+++ lib/app.js
@@ -1,0 +1,3 @@
+﻿var Nightmare = require("./nightmare");
+require("./cookies");
+module.exports = Nightmare;
