+++ lib/nightmare.js
@@ -640,7 +640,6 @@ class Nightmare {
         child.on('destroyed', function () { eventLog('destroyed', JSON.stringify(Array.from(arguments))); });
 
         self.child = child;
-        self.cookies = new Cookies(this);
 
         return co(function* () {
             self.initializeNamespaces();
@@ -699,7 +698,9 @@ class Nightmare {
             let p = this._invokeRunnerOperation(operationParams, electronActionName, String(Nightmare._electronActions[electronActionName]), id);
             promises.push(p);
         }
-        return Promise.all(promises);
+        return Promise.all(promises).then(function () {
+            debug('.initializeElectronActions() completed.');
+        });
     }
 
     /**
@@ -1218,73 +1219,6 @@ class Nightmare {
         }
 
     }
-}
-
-class Cookies {
-
-    constructor(nightmare) {
-        this._nightmare = nightmare;
-        this.child = this._nightmare.child;
-    }
-
-    /**
-    * Get a cookie
-    */
-    get(name) {
-        debug('cookies.get()')
-        let query = {}
-
-        if (_.isObject(name))
-            query = name;
-        else
-            query.name = name;
-
-        return this._nightmare._invokeRunnerOperation("cookie.get", query);
-    };
-
-    /**
-     * Set a cookie
-     */
-    set(name, value) {
-        debug('cookies.set()')
-
-        let cookies = [];
-        if (_.isArray(name))
-            cookies = name;
-        else if (_.isObject(name))
-            cookies.push(name);
-        else cookies.push({
-            name: name,
-            value: value
-        });
-
-        if (cookies.length === 0)
-            return this._nightmare._noop();
-
-        return this._nightmare._invokeRunnerOperation("cookie.set", cookies);
-    };
-
-    /**
-     * Clear a cookie
-     */
-    clear(name, url) {
-        debug('cookies.clear()');
-
-        let cookies = [];
-        if (_.isArray(name))
-            cookies = name;
-        else if (_.isObject(name))
-            cookies.push(name);
-        else cookies.push({
-            name: name,
-            url: url
-        });
-
-        if (cookies.length == 0)
-            return this._nightmare._noop();
-
-        return this._nightmare._invokeRunnerOperation("cookie.clear", cookies);
-    };
 }
 
 Nightmare._namespaces = [];
