+++ lib/runner.js
@@ -504,66 +504,6 @@ app.on('ready', function () {
     });
 
     /**
-     * Get cookies
-     */
-
-    parent.on('cookie.get', function (query) {
-        var details = _.assign({}, {
-            url: win.webContents.getURL(),
-        }, query)
-
-        parent.emit('log', 'getting cookie: ' + JSON.stringify(details))
-        win.webContents.session.cookies.get(details, function (err, cookies) {
-            if (err) return parent.emit('cookie.get', err);
-            parent.emit('cookie.get', {
-                result: details.name ? cookies[0] : cookies
-            });
-        })
-    })
-
-    /**
-     * Set cookies
-     */
-
-    parent.on('cookie.set', function (cookies) {
-        var pending = cookies.length;
-        for (var cookie of cookies) {
-            var details = _.assign({}, {
-                url: win.webContents.getURL()
-            }, cookie)
-
-            parent.emit('log', 'setting cookie: ' + JSON.stringify(details))
-            win.webContents.session.cookies.set(details, function (err) {
-                if (err) parent.emit('cookie.set', {
-                    error: err
-                });
-                else if (!--pending) parent.emit('cookie.set')
-            })
-        }
-    })
-
-    /**
-     * Clear cookie
-     */
-    parent.on('cookie.clear', function (cookies) {
-        var pending = cookies.length;
-        parent.emit('log', 'listing params', cookies);
-
-        for (var cookie of cookies) {
-            let url = cookie.url || win.webContents.getURL();
-            let name = cookie.name;
-
-            parent.emit('log', 'clearing cookie: ' + JSON.stringify(cookie))
-            win.webContents.session.cookies.remove(url, name, function (err) {
-                if (err) parent.emit('cookie.clear', {
-                    error: err,
-                });
-                else if (!--pending) parent.emit('cookie.clear')
-            });
-        }
-    });
-
-    /**
      * Wait until Finish Load
      */
     parent.on("waitUntilFinishLoad", function () {
